Reorganizing the llm examples directory
