+++ examples/llm/cli.py
@@ -13,6 +13,7 @@
 # limitations under the License.
 
 import logging
+import time
 
 import click
 
@@ -22,7 +23,6 @@ from llm.rag import run as run_rag
 from llm.vdb_upload import run as run_vdb_upload
 from morpheus.cli.utils import get_log_levels
 from morpheus.cli.utils import parse_log_level
-from morpheus.utils.logger import configure_logging
 
 
 @click.group(name="morpheus_llm", chain=False, invoke_without_command=True, no_args_is_help=True)
@@ -31,9 +31,22 @@ from morpheus.utils.logger import configure_logging
               type=click.Choice(get_log_levels(), case_sensitive=False),
               callback=parse_log_level,
               help="Specify the logging level to use.")
+@click.option('--use_cpp',
+              default=True,
+              type=bool,
+              help=("Whether or not to use C++ node and message types or to prefer python. "
+                    "Only use as a last resort if bugs are encountered"))
 @click.version_option()
-def cli(log_level: int):
+@click.pass_context
+def cli(ctx: click.Context, log_level: int, use_cpp: bool):
     """Main entrypoint for the LLM Examples"""
+
+    from morpheus.config import CppConfig
+    from morpheus.utils.logger import configure_logging
+
+    ctx_dict = ctx.ensure_object(dict)
+
+    CppConfig.set_should_use_cpp(use_cpp)
 
     # Configure the logging
     configure_logging(log_level=log_level)
@@ -45,8 +58,27 @@ def cli(log_level: int):
     # Set the parent logger for all of the llm examples to use morpheus so we can take advantage of configure_logging
     logger.parent = morpheus_logger
 
+    ctx_dict["start_time"] = time.time()
+
 
 cli.add_command(run_vdb_upload, name='vdb_upload')
 cli.add_command(run_completion, name='completion')
 cli.add_command(run_rag, name='rag')
 cli.add_command(run_agents, name='agents')
+
+
+@cli.result_callback()
+@click.pass_context
+def after_pipeline(ctx: click.Context, pipeline_start_time: float, *_, **__):
+    logger = logging.getLogger(__name__)
+
+    end_time = time.time()
+
+    ctx_dict = ctx.ensure_object(dict)
+
+    start_time = ctx_dict["start_time"]
+
+    logger.info("Total time: %.2f sec", end_time - start_time)
+
+    if (pipeline_start_time is not None):
+        logger.info("Pipeline runtime: %.2f sec", end_time - pipeline_start_time)
