+++ examples/llm/rag/run.py
@@ -13,129 +13,10 @@
 # limitations under the License.
 import logging
 import os
-import time
 
 import click
 
 logger = logging.getLogger(f"morpheus.{__name__}")
-
-
-def _build_embeddings(model_name: str):
-    from langchain.embeddings import HuggingFaceEmbeddings
-
-    model_name = f"sentence-transformers/{model_name}"
-
-    model_kwargs = {'device': 'cuda'}
-    encode_kwargs = {
-        # 'normalize_embeddings': True, # set True to compute cosine similarity
-        "batch_size": 100,
-    }
-
-    embeddings = HuggingFaceEmbeddings(model_name=model_name, model_kwargs=model_kwargs, encode_kwargs=encode_kwargs)
-
-    return embeddings
-
-
-def _build_milvus_service():
-    import pymilvus
-
-    from morpheus.service.milvus_vector_db_service import MilvusVectorDBService
-    from morpheus.utils.vector_db_service_utils import VectorDBServiceFactory
-
-    milvus_resource_kwargs = {
-        "index_conf": {
-            "field_name": "embedding",
-            "metric_type": "L2",
-            "index_type": "HNSW",
-            "params": {
-                "M": 8,
-                "efConstruction": 64,
-            },
-        },
-        "schema_conf": {
-            "enable_dynamic_field": True,
-            "schema_fields": [
-                pymilvus.FieldSchema(name="id",
-                                     dtype=pymilvus.DataType.INT64,
-                                     description="Primary key for the collection",
-                                     is_primary=True,
-                                     auto_id=True).to_dict(),
-                pymilvus.FieldSchema(name="title",
-                                     dtype=pymilvus.DataType.VARCHAR,
-                                     description="The title of the RSS Page",
-                                     max_length=65_535).to_dict(),
-                pymilvus.FieldSchema(name="link",
-                                     dtype=pymilvus.DataType.VARCHAR,
-                                     description="The URL of the RSS Page",
-                                     max_length=65_535).to_dict(),
-                pymilvus.FieldSchema(name="summary",
-                                     dtype=pymilvus.DataType.VARCHAR,
-                                     description="The summary of the RSS Page",
-                                     max_length=65_535).to_dict(),
-                pymilvus.FieldSchema(name="page_content",
-                                     dtype=pymilvus.DataType.VARCHAR,
-                                     description="A chunk of text from the RSS Page",
-                                     max_length=65_535).to_dict(),
-                pymilvus.FieldSchema(name="embedding",
-                                     dtype=pymilvus.DataType.FLOAT_VECTOR,
-                                     description="Embedding vectors",
-                                     dim=384).to_dict(),
-            ],
-            "description": "Test collection schema"
-        }
-    }
-
-    vdb_service: MilvusVectorDBService = VectorDBServiceFactory.create_instance("milvus",
-                                                                                uri="http://localhost:19530",
-                                                                                **milvus_resource_kwargs)
-
-    return vdb_service.load_resource("Arxiv")
-
-
-def _build_llm_service(model_name: str):
-
-    from ..common.nemo_llm_service import NeMoLLMService
-
-    llm_service = NeMoLLMService()
-
-    return llm_service.get_client(model_name=model_name, temperature=0.0)
-
-
-def _build_engine(model_name: str):
-
-    from morpheus.llm import LLMEngine
-
-    from ..common.extracter_node import ExtracterNode
-    from ..common.rag_node import RAGNode
-    from ..common.simple_task_handler import SimpleTaskHandler
-
-    engine = LLMEngine()
-
-    engine.add_node("extracter", node=ExtracterNode())
-
-    prompt = """You are a helpful assistant. Given the following background information:\n
-{% for c in contexts -%}
-Title: {{ c.title }}
-Summary: {{ c.summary }}
-Text: {{ c.page_content }}
-{% endfor %}
-
-Please answer the following question: \n{{ query }}"""
-
-    vector_service = _build_milvus_service()
-    embeddings = _build_embeddings("all-MiniLM-L6-v2")
-    llm_service = _build_llm_service(model_name)
-
-    engine.add_node("rag",
-                    inputs=["/extracter"],
-                    node=RAGNode(prompt=prompt,
-                                 vdb_service=vector_service,
-                                 embedding=embeddings.embed_documents,
-                                 llm_client=llm_service))
-
-    engine.add_task_handler(inputs=["/extracter"], handler=SimpleTaskHandler())
-
-    return engine
 
 
 @click.group(name=__name__)
@@ -170,64 +51,54 @@ def run():
     default='gpt-43b-002',
     help="The name of the model that is deployed on Triton server",
 )
-def pipeline(
-    num_threads,
-    pipeline_batch_size,
-    model_max_batch_size,
-    model_name,
-):
+@click.option(
+    "--repeat_count",
+    default=1,
+    type=click.IntRange(min=1),
+    help="Number of times to repeat the input query. Useful for testing performance.",
+)
+def pipeline(**kwargs):
 
-    import cudf
+    from .standalone_pipeline import standalone
 
-    from morpheus.config import Config
-    from morpheus.config import CppConfig
-    from morpheus.config import PipelineModes
-    from morpheus.messages import ControlMessage
-    from morpheus.pipeline.linear_pipeline import LinearPipeline
-    from morpheus.stages.general.monitor_stage import MonitorStage
-    from morpheus.stages.input.in_memory_source_stage import InMemorySourceStage
-    from morpheus.stages.output.in_memory_sink_stage import InMemorySinkStage
-    from morpheus.stages.preprocess.deserialize_stage import DeserializeStage
+    return standalone(**kwargs)
 
-    from ..common.llm_engine_stage import LLMEngineStage
 
-    CppConfig.set_should_use_cpp(False)
+@run.command()
+@click.option(
+    "--num_threads",
+    default=os.cpu_count(),
+    type=click.IntRange(min=1),
+    help="Number of internal pipeline threads to use",
+)
+@click.option(
+    "--pipeline_batch_size",
+    default=1024,
+    type=click.IntRange(min=1),
+    help=("Internal batch size for the pipeline. Can be much larger than the model batch size. "
+          "Also used for Kafka consumers"),
+)
+@click.option(
+    "--model_max_batch_size",
+    default=64,
+    type=click.IntRange(min=1),
+    help="Max batch size to use for the model",
+)
+@click.option(
+    "--embedding_size",
+    default=384,
+    type=click.IntRange(min=1),
+    help="The output size of the embedding calculation. Depends on the model supplied by --model_name",
+)
+@click.option(
+    "--model_name",
+    required=True,
+    type=str,
+    default='gpt-43b-002',
+    help="The name of the model that is deployed on Triton server",
+)
+def persistant(**kwargs):
 
-    config = Config()
-    config.mode = PipelineModes.OTHER
+    from .persistant_pipeline import pipeline as _pipeline
 
-    # Below properties are specified by the command line
-    config.num_threads = num_threads
-    config.pipeline_batch_size = pipeline_batch_size
-    config.model_max_batch_size = model_max_batch_size
-    config.mode = PipelineModes.NLP
-    config.edge_buffer_size = 128
-
-    source_dfs = [cudf.DataFrame({"questions": ["Tell me a story about your best friend."] * 5})]
-
-    completion_task = {"task_type": "completion", "task_dict": {"input_keys": ["questions"], }}
-
-    pipe = LinearPipeline(config)
-
-    pipe.set_source(InMemorySourceStage(config, dataframes=source_dfs, repeat=10))
-
-    pipe.add_stage(
-        DeserializeStage(config, message_type=ControlMessage, task_type="llm_engine", task_payload=completion_task))
-
-    pipe.add_stage(MonitorStage(config, description="Source rate", unit='questions'))
-
-    pipe.add_stage(LLMEngineStage(config, engine=_build_engine(model_name=model_name)))
-
-    sink = pipe.add_stage(InMemorySinkStage(config))
-
-    pipe.add_stage(MonitorStage(config, description="Upload rate", unit="events", delayed_start=True))
-
-    start_time = time.time()
-
-    pipe.run()
-
-    duration = time.time() - start_time
-
-    print("Got messages: ", sink.get_messages())
-
-    print(f"Total duration: {duration:.2f} seconds")
+    return _pipeline(**kwargs)
