+++ examples/llm/rag/standalone_pipeline.py
@@ -1,0 +1,196 @@
+# Copyright (c) 2023, NVIDIA CORPORATION.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+import logging
+import time
+
+import pymilvus
+from langchain.embeddings import HuggingFaceEmbeddings
+
+import cudf
+
+from morpheus.config import Config
+from morpheus.config import PipelineModes
+from morpheus.llm import LLMEngine
+from morpheus.messages import ControlMessage
+from morpheus.pipeline.linear_pipeline import LinearPipeline
+from morpheus.service.milvus_vector_db_service import MilvusVectorDBService
+from morpheus.stages.general.monitor_stage import MonitorStage
+from morpheus.stages.input.in_memory_source_stage import InMemorySourceStage
+from morpheus.stages.output.in_memory_sink_stage import InMemorySinkStage
+from morpheus.stages.preprocess.deserialize_stage import DeserializeStage
+from morpheus.utils.vector_db_service_utils import VectorDBServiceFactory
+
+from ..common.extracter_node import ExtracterNode
+from ..common.llm_engine_stage import LLMEngineStage
+from ..common.nemo_llm_service import NeMoLLMService
+from ..common.rag_node import RAGNode
+from ..common.simple_task_handler import SimpleTaskHandler
+
+logger = logging.getLogger(f"morpheus.{__name__}")
+
+
+def _build_embeddings(model_name: str):
+
+    model_name = f"sentence-transformers/{model_name}"
+
+    model_kwargs = {'device': 'cuda'}
+    encode_kwargs = {
+        # 'normalize_embeddings': True, # set True to compute cosine similarity
+        "batch_size": 100,
+    }
+
+    embeddings = HuggingFaceEmbeddings(model_name=model_name, model_kwargs=model_kwargs, encode_kwargs=encode_kwargs)
+
+    return embeddings
+
+
+def _build_milvus_service():
+
+    milvus_resource_kwargs = {
+        "index_conf": {
+            "field_name": "embedding",
+            "metric_type": "L2",
+            "index_type": "HNSW",
+            "params": {
+                "M": 8,
+                "efConstruction": 64,
+            },
+        },
+        "schema_conf": {
+            "enable_dynamic_field": True,
+            "schema_fields": [
+                pymilvus.FieldSchema(name="id",
+                                     dtype=pymilvus.DataType.INT64,
+                                     description="Primary key for the collection",
+                                     is_primary=True,
+                                     auto_id=True).to_dict(),
+                pymilvus.FieldSchema(name="title",
+                                     dtype=pymilvus.DataType.VARCHAR,
+                                     description="The title of the RSS Page",
+                                     max_length=65_535).to_dict(),
+                pymilvus.FieldSchema(name="link",
+                                     dtype=pymilvus.DataType.VARCHAR,
+                                     description="The URL of the RSS Page",
+                                     max_length=65_535).to_dict(),
+                pymilvus.FieldSchema(name="summary",
+                                     dtype=pymilvus.DataType.VARCHAR,
+                                     description="The summary of the RSS Page",
+                                     max_length=65_535).to_dict(),
+                pymilvus.FieldSchema(name="page_content",
+                                     dtype=pymilvus.DataType.VARCHAR,
+                                     description="A chunk of text from the RSS Page",
+                                     max_length=65_535).to_dict(),
+                pymilvus.FieldSchema(name="embedding",
+                                     dtype=pymilvus.DataType.FLOAT_VECTOR,
+                                     description="Embedding vectors",
+                                     dim=384).to_dict(),
+            ],
+            "description": "Test collection schema"
+        }
+    }
+
+    vdb_service: MilvusVectorDBService = VectorDBServiceFactory.create_instance("milvus",
+                                                                                uri="http://localhost:19530",
+                                                                                **milvus_resource_kwargs)
+
+    return vdb_service.load_resource("Arxiv")
+
+
+def _build_llm_service(model_name: str):
+
+    llm_service = NeMoLLMService()
+
+    return llm_service.get_client(model_name=model_name, temperature=0.5, tokens_to_generate=200)
+
+
+def _build_engine(model_name: str):
+
+    engine = LLMEngine()
+
+    engine.add_node("extracter", node=ExtracterNode())
+
+    prompt = """You are a helpful assistant. Given the following background information:\n
+{% for c in contexts -%}
+Title: {{ c.title }}
+Summary: {{ c.summary }}
+Text: {{ c.page_content }}
+{% endfor %}
+
+Please answer the following question: \n{{ query }}"""
+
+    vector_service = _build_milvus_service()
+    embeddings = _build_embeddings("all-MiniLM-L6-v2")
+    llm_service = _build_llm_service(model_name)
+
+    # Async wrapper around embeddings
+    async def calc_embeddings(texts: list[str]) -> list[list[float]]:
+        return embeddings.embed_documents(texts)
+
+    engine.add_node("rag",
+                    inputs=["/extracter"],
+                    node=RAGNode(prompt=prompt,
+                                 vdb_service=vector_service,
+                                 embedding=calc_embeddings,
+                                 llm_client=llm_service))
+
+    engine.add_task_handler(inputs=["/rag"], handler=SimpleTaskHandler())
+
+    return engine
+
+
+def standalone(
+    num_threads,
+    pipeline_batch_size,
+    model_max_batch_size,
+    model_name,
+    repeat_count,
+):
+    config = Config()
+    config.mode = PipelineModes.OTHER
+
+    # Below properties are specified by the command line
+    config.num_threads = num_threads
+    config.pipeline_batch_size = pipeline_batch_size
+    config.model_max_batch_size = model_max_batch_size
+    config.mode = PipelineModes.NLP
+    config.edge_buffer_size = 128
+
+    source_dfs = [
+        cudf.DataFrame({"questions": ["What are some new attacks discovered in the cyber security industry?."] * 5})
+    ]
+
+    completion_task = {"task_type": "completion", "task_dict": {"input_keys": ["questions"], }}
+
+    pipe = LinearPipeline(config)
+
+    pipe.set_source(InMemorySourceStage(config, dataframes=source_dfs, repeat=repeat_count))
+
+    pipe.add_stage(
+        DeserializeStage(config, message_type=ControlMessage, task_type="llm_engine", task_payload=completion_task))
+
+    pipe.add_stage(MonitorStage(config, description="Source rate", unit='questions'))
+
+    pipe.add_stage(LLMEngineStage(config, engine=_build_engine(model_name=model_name)))
+
+    sink = pipe.add_stage(InMemorySinkStage(config))
+
+    pipe.add_stage(MonitorStage(config, description="Response rate", unit="responses", delayed_start=True))
+
+    start_time = time.time()
+
+    pipe.run()
+
+    logger.info("Pipeline complete. Received %s responses", len(sink.get_messages()))
+
+    return start_time
