+++ examples/llm/vdb_upload/common.py
@@ -1,0 +1,103 @@
+# Copyright (c) 2023, NVIDIA CORPORATION.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+import logging
+
+import pymilvus
+
+logger = logging.getLogger(f"morpheus.{__name__}")
+
+
+def build_milvus_config(embedding_size: int):
+
+    milvus_resource_kwargs = {
+        "index_conf": {
+            "field_name": "embedding",
+            "metric_type": "L2",
+            "index_type": "HNSW",
+            "params": {
+                "M": 8,
+                "efConstruction": 64,
+            },
+        },
+        "schema_conf": {
+            "enable_dynamic_field": True,
+            "schema_fields": [
+                pymilvus.FieldSchema(name="id",
+                                     dtype=pymilvus.DataType.INT64,
+                                     description="Primary key for the collection",
+                                     is_primary=True,
+                                     auto_id=True).to_dict(),
+                pymilvus.FieldSchema(name="title",
+                                     dtype=pymilvus.DataType.VARCHAR,
+                                     description="The title of the RSS Page",
+                                     max_length=65_535).to_dict(),
+                pymilvus.FieldSchema(name="link",
+                                     dtype=pymilvus.DataType.VARCHAR,
+                                     description="The URL of the RSS Page",
+                                     max_length=65_535).to_dict(),
+                pymilvus.FieldSchema(name="summary",
+                                     dtype=pymilvus.DataType.VARCHAR,
+                                     description="The summary of the RSS Page",
+                                     max_length=65_535).to_dict(),
+                pymilvus.FieldSchema(name="page_content",
+                                     dtype=pymilvus.DataType.VARCHAR,
+                                     description="A chunk of text from the RSS Page",
+                                     max_length=65_535).to_dict(),
+                pymilvus.FieldSchema(name="embedding",
+                                     dtype=pymilvus.DataType.FLOAT_VECTOR,
+                                     description="Embedding vectors",
+                                     dim=embedding_size).to_dict(),
+            ],
+            "description": "Test collection schema"
+        }
+    }
+
+    return milvus_resource_kwargs
+
+
+def build_rss_urls():
+    return [
+        "https://www.theregister.com/security/headlines.atom",
+        "https://isc.sans.edu/dailypodcast.xml",
+        "https://threatpost.com/feed/",
+        "http://feeds.feedburner.com/TheHackersNews?format=xml",
+        "https://www.bleepingcomputer.com/feed/",
+        "https://therecord.media/feed/",
+        "https://blog.badsectorlabs.com/feeds/all.atom.xml",
+        "https://krebsonsecurity.com/feed/",
+        "https://www.darkreading.com/rss_simple.asp",
+        "https://blog.malwarebytes.com/feed/",
+        "https://msrc.microsoft.com/blog/feed",
+        "https://securelist.com/feed",
+        "https://www.crowdstrike.com/blog/feed/",
+        "https://threatconnect.com/blog/rss/",
+        "https://news.sophos.com/en-us/feed/",
+        "https://www.us-cert.gov/ncas/current-activity.xml",
+        "https://www.csoonline.com/feed",
+        "https://www.cyberscoop.com/feed",
+        "https://research.checkpoint.com/feed",
+        "https://feeds.fortinet.com/fortinet/blog/threat-research",
+        "https://www.mcafee.com/blogs/rss",
+        "https://www.digitalshadows.com/blog-and-research/rss.xml",
+        "https://www.nist.gov/news-events/cybersecurity/rss.xml",
+        "https://www.sentinelone.com/blog/rss/",
+        "https://www.bitdefender.com/blog/api/rss/labs/",
+        "https://www.welivesecurity.com/feed/",
+        "https://unit42.paloaltonetworks.com/feed/",
+        "https://mandiant.com/resources/blog/rss.xml",
+        "https://www.wired.com/feed/category/security/latest/rss",
+        "https://www.wired.com/feed/tag/ai/latest/rss",
+        "https://blog.google/threat-analysis-group/rss/",
+        "https://intezer.com/feed/",
+    ]
