+++ examples/llm/vdb_upload/langchain.py
@@ -1,0 +1,60 @@
+# Copyright (c) 2023, NVIDIA CORPORATION.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+import logging
+import pickle
+
+from examples.llm.vdb_upload.common import build_rss_urls
+from langchain.document_loaders.rss import RSSFeedLoader
+from langchain.embeddings.huggingface import HuggingFaceEmbeddings
+from langchain.text_splitter import RecursiveCharacterTextSplitter
+from langchain.vectorstores.milvus import Milvus
+
+from morpheus.utils.logging_timer import log_time
+
+logger = logging.getLogger(f"morpheus.{__name__}")
+
+
+def chain(model_name, save_cache):
+
+    with log_time(msg="Seeding with chain took {duration} ms. {rate_per_sec} docs/sec", log_fn=logger.debug) as log:
+
+        loader = RSSFeedLoader(urls=build_rss_urls())
+
+        documents = loader.load()
+
+        if (save_cache is not None):
+            with open(save_cache, "wb") as f:
+                pickle.dump(documents, f)
+
+        text_splitter = RecursiveCharacterTextSplitter(chunk_size=200, chunk_overlap=20, length_function=len)
+
+        documents = text_splitter.split_documents(documents)
+
+        log.count = len(documents)
+
+        logger.info("Loaded %s documents", len(documents))
+
+        embeddings = HuggingFaceEmbeddings(
+            model_name=model_name,
+            model_kwargs={'device': 'cuda'},
+            encode_kwargs={
+                # 'normalize_embeddings': True, # set True to compute cosine similarity
+                "batch_size": 100,
+            })
+
+        with log_time(msg="Adding to Milvus took {duration} ms. Doc count: {count}. {rate_per_sec} docs/sec",
+                      count=log.count,
+                      log_fn=logger.debug):
+
+            Milvus.from_documents(documents, embeddings, collection_name="LangChain", drop_old=True)
